adapt to new tt interface

+++ media/server/main/include/ITextTrackAccessor.h
@@ -58,7 +58,7 @@ public:
     virtual bool play(uint32_t sessionId) = 0;
     virtual bool mute(uint32_t sessionId, bool mute) = 0;
     virtual bool setPosition(uint32_t sessionId, uint64_t mediaTimestampMs) = 0;
-    virtual bool sendData(uint32_t sessionId, const std::string &data, DataType datatype, int32_t displayOffsetMs = 0) = 0;
+    virtual bool sendData(uint32_t sessionId, const std::string &data, DataType datatype, int64_t displayOffsetMs = 0) = 0;
     virtual bool setSessionWebVTTSelection(uint32_t sessionId) = 0;
     virtual bool setSessionTTMLSelection(uint32_t sessionId) = 0;
     virtual bool setSessionCCSelection(uint32_t sessionId, const std::string &service) = 0;
