+++ wrappers/include/TextTrackWrapper.h
@@ -43,7 +43,7 @@ public:
     std::uint32_t resetSession(std::uint32_t sessionId) const override;
     std::uint32_t sendSessionTimestamp(std::uint32_t sessionId, std::uint64_t mediaTimestampMs) const override;
     std::uint32_t sendSessionData(std::uint32_t sessionId, ITextTrackWrapper::DataType type,
-                                  std::int32_t displayOffsetMs, const std::string &data) const override;
+                                  std::int64_t displayOffsetMs, const std::string &data) const override;
     std::uint32_t setSessionWebVTTSelection(std::uint32_t sessionId) const override;
     std::uint32_t setSessionTTMLSelection(std::uint32_t sessionId) const override;
     std::uint32_t setSessionClosedCaptionsService(std::uint32_t sessionId, const std::string &service) const override;
